fix(techdebt): float64-safe sparse conversion + general PSD test + doc counts/version
- nn/hodge.py: sparse_scipy_to_torch no longer casts COO data through an
  intermediate np.float32 before re-casting to the requested dtype. float64
  callers silently lost ~1e-12 of precision (elite-code §1.3). Now casts the
  scipy data directly to the requested dtype (float64 in -> float64 out, exact).

- tests/test_hodge.py: add test_sparse_scipy_to_torch_float64_no_precision_loss.
  Uses a value not representable in float32 (1.0 + 1e-12) and asserts exact
  equality (np.testing.assert_array_equal) so the test fails if a float32
  intermediate is reintroduced. Verified to fail under the old code path.

- tests/test_complexes.py: strengthen test_hodge_laplacian_symmetric_psd from a
  single near-degenerate facet (0,1,2) to a parametrized check that also covers a
  non-degenerate general complex (two triangles sharing edge (1,2): 4 vertices,
  5 edges, 2 faces), per the research-math proof bar (test the general case).
  The degenerate triangle case is kept, not replaced.

- CONTRIBUTING.md: test count 491 -> 497 (canonical `grep -c "def test_"` = 497,
  matching README/STATUS/CHANGELOG/CLAIMS_TO_EVIDENCE).

- LIMITATIONS.md: document-identity version bumped to current package version
  0.0.3 (title, lead, and the "How to read the version" signal, which now reads
  0.0.3 not v0.0.1-alpha). Scientific-milestone / release-gate / historical
  v0.0.1 references are intentionally left unchanged.

+++ topogeoml/nn/hodge.py
@@ -78,7 +78,9 @@ def sparse_scipy_to_torch(
     """Convert a scipy sparse matrix to a torch sparse COO tensor."""
     coo = matrix.tocoo()
     indices = torch.from_numpy(np.vstack([coo.row, coo.col]).astype(np.int64))
-    values = torch.from_numpy(coo.data.astype(np.float32)).to(dtype)
+    # Cast directly to the requested dtype — no float32 intermediate, which would
+    # silently truncate precision for float64 callers (elite-code §1.3).
+    values = torch.from_numpy(coo.data).to(dtype)
     shape = torch.Size(coo.shape)
     tensor = torch.sparse_coo_tensor(indices, values, shape).coalesce()
     if device is not None:
